requirements.txt
Flask~=2.2.2
Gunicorn~=20.1.0
requests~=2.28.1
google~=3.0.0
google-api-python-client~=2.65.0
python-docx~=1.1.0